# Switch to sync.Map to improve performance when under lots of load