--- a/main_test.go
+++ b/main_test.go
@@ -123,10 +123,8 @@ func TestProcessLine(t *testing.T) {
 	c.ch <- nil
 	for _, k := range testCases {
 		originalName := strings.Split(k.line, " ")[0]
-		data, ok := c.samples.Load(originalName)
-		assert.Equal(t, ok, true)
-		sample, ok := data.(*graphiteSample)
-		assert.Equal(t, ok, true)
+		data, _ := c.samples.Load(originalName)
+		sample, _ := data.(*graphiteSample)
 
 		if k.willFail {
 			assert.Nil(t, sample, "Found %s", k.name)