Part of https://github.com/prometheus/graphite_exporter/issues/66. The idea behind this change is to lock the map per key but not per the whole instance.
Also, switch a few methods so that there would be no copying of `sync.Map`. That might have improved the performance as well because there are less map copying.

After:
[profilev2.pdf](https://github.com/prometheus/graphite_exporter/files/2653061/profilev2.pdf)

Before:
[profiling_output.pdf](https://github.com/prometheus/graphite_exporter/files/2653062/profiling_output.pdf)

I don't have the graphs for lengths of how long it took to download `/metrics` with ~500k metrics but now it seems that the lengths are much more consistent even though they still take a long time. It used to jump from ~2s. to ~30s. With this change it consistently takes around ~10s. which would you expect to happen since the stress testing program pushes new metrics each 10 seconds.