@matthiasr 2018-12-11
Hmm, I'm not sure going from 2s to 10s in the base case is really an improvement, especially since 10s is already the default scrape timeout.

Is locking per key beneficial for the scrape case that needs to look at all the values anyway?

What if we change the Mutex to an RWMutex but use a regular map? Would that help?

Could you also post flame graphs? I find them easier to read for "where is time being spent".