@GiedriusS 2018-12-15
My point was that having a consistent ~10s. delay is better than having a delay that jumps anywhere from 2 s. to even 30 s. because the user then would know for sure that something is wrong. Even I myself got tricked into believing that everything is OK because I refreshed when that scrape had happened successfully due to pure luck. `sync.Map` already uses a Mutex underneath that is only held when necessary: https://golang.org/src/sync/map.go?s=1149:2596#L17 so essentially we would be reinventing the wheel here, IMHO. I will post the flame graphs when I can but please consider this again.