@matthiasr 2018-12-18
I agree that the current status is not great, but if the baseline is 2 seconds, I would rather not 5-times that. This points at some new overhead that we are introducing and that I would like to avoid as much as possible.

I am not claiming that an RWMutex would do better but I think it's worth trying. I don't think moving from a Mutex to an RWMutex is reinventing anything.

Could you share your test program?