@matthiasr 2019-10-07
I'm going to close this because the mainline code has changed a ton, including significant performance improvements from the `client_golang` update.

Please benchmark again, and open a new PR if we still have significant potential for improvement.